+++ RELEASE-NOTES.md
@@ -5,6 +5,7 @@
 ### Code Changes
 
 - Adds the `data_exported_at` column to the `actions` table in the database.
+- Adds the `ETL_SERVER_URL` environment variable to the backend in order to configure the ETL server URL.
 
 ## v3.2.1